% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.r
\name{alpha}
\alias{alpha}
\title{Change colour opacity}
\usage{
alpha(x, alpha)
}
\arguments{
\item{x}{vector of colours, in any format that \code{col2rgb()} accepts}

\item{alpha}{factor modifying the opacity alpha; typically in \verb{[0,1]}}
}
\value{
a color vector of the same length as \code{x} with requested alpha value
}
\description{
Change opacity of colours by modifying the alpha value (typically \verb{[0,1]}).
}
\details{
This is a shorthand for \code{adjustcolor(x, alpha.f = alpha)} which offers more knobs to modify.
Adding alpha to colours is useful for overploting multiple elements,
such as plotting multiple points with similar values.
}
\examples{
# These are identical:
grDevices::adjustcolor("red", alpha.f = 0.5)
alpha("red", 0.5)

# Unlike adjustcolor, alpha preserves names
col = c("red", "blue", "green") |> setNames(c("foo", "bar", "baz"))
alpha(col, 0.5)

}
\seealso{
\code{\link[grDevices:adjustcolor]{grDevices::adjustcolor()}}
}
