% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.r
\name{blend}
\alias{blend}
\title{Blend colors in RGB space}
\usage{
blend(x, y = NULL, ratio = 1)
}
\arguments{
\item{x, y}{a vector of colours specified either by name or in a exadecimal form, see details}

\item{ratio}{a ratio in which colours are blended}
}
\value{
If \code{y = NULL}, a single value resulting from blending elements of \code{x}.
If both \code{x} and \code{y} are specified, then a vector of the same length as the longer of \code{x} and \code{y}.
}
\description{
Blend together two or more colours in RGB space by averaging the colours in specified ratio.
}
\details{
If only \code{x} is specified, all colours in \code{x} are averaged and a single colour is returned.
If both \code{x} and \code{y} are specified, they are averaged element-wise and vector of the same
length as the longer of \code{x} and \code{y} is returned.
}
\examples{
# Mix two colors in 1:1 ratio:
blend("black", "white")
blend(c("black", "white"))

# Mix two colours in 5:1 ratio:
blend("black", "white", 5)
blend(c("black", "white"), ratio = 5)

# Ratio is extended with 1 as required:
blend(c("red", "green", "blue"))
blend(c("red", "green", "blue"), ratio = 1)
blend(c("red", "green", "blue"), ratio = c(1, 1))
blend(c("red", "green", "blue", ratio = c(1, 1, 1)))

# Ratio 1:1 and 2:2 etc. are identical
blend("black", "white", c(1, 1))
blend("black", "white", c(2, 2))

# Make a vector of colours a little lighter
col = palette.colors(5, "Set 1")
blend(col, "white", c(2, 1))
blend("white", col, c(1, 2))

}
\seealso{
\link[grDevices:adjustcolor]{grDevices::adjustcolor} for adjusting a vector of colours.
}
