% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_cases.r
\name{complete_cases}
\alias{complete_cases}
\title{Find complete cases}
\usage{
complete_cases(x, select = NULL)
}
\arguments{
\item{x}{a matrix or a data.frame}

\item{select}{a numeric vector of indices or a character vector of column names,
the presence of missing values is checked only in these columns, other columns are ignored}
}
\value{
a matrix or a data.frame with no missing values in selected columns
}
\description{
Find complete cases in a matrix or a data.fame, i.e., rows with no missing values.
}
\details{
This function is a user-friendly wrapper for \code{\link[stats:complete.cases]{stats::complete.cases()}}, which returns
a logical vector.
}
\examples{
d = data.frame(
  "foo" = c(1, NA, 3),
  "bar" = c(NA, 2, 3)
  )

# Following are identical
complete_cases(d)
d[stats::complete.cases(d),]

# Consider only first column
complete_cases(d, "foo")
complete_cases(d, 1)
d[stats::complete.cases(d[1]),]

}
\seealso{
\code{\link[stats:complete.cases]{stats::complete.cases()}} for a similar function returning a logical indicator,
\code{\link[base:NA]{base::is.na()}} for checking if the vector or a matrix contains \code{NA},
\code{\link[stats:na.fail]{stats::na.omit()}} for a generic used to remove \code{NA}s.
}
