% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir.r
\name{mkdir}
\alias{mkdir}
\title{Make directory}
\usage{
mkdir(x, recursive = TRUE, showWarnings = TRUE, mode = "0777")
}
\arguments{
\item{x}{one or more paths}

\item{recursive}{logical, should elements of the paths be created? If set to \code{FALSE} and the
directory \code{foo} doesn't exist, then \verb{mkdir("foo/bar, recursive = FALSE)} would fail.}

\item{showWarnings}{logical, should the warning on failure be shown?}

\item{mode}{the mode to be used on Unix-alike, see \code{?dir.create} for more information.}
}
\value{
A logical vector indicating if the operation succeeded for each of the directories
or if the directory already exists.
}
\description{
Make one or more directories.
}
\details{
This function provides a different interface for \code{dir.create} that is more similar to
the Unix command \code{mkdir} (specifically \code{mkdir -p}). The two major differences
being a vectorization over \code{x} so unlike \code{dir.create} this functions accepts a vector.
The other difference being a test for the directory existence so a directory is created
only if it doesn't already exists.
}
\examples{
\dontrun{
# create multiple directories at once
mkdir(c("foo", "bar"))

# Following are identical:
if(!dir.exist("foo/bar")
   dir.create("foo/bar", recursive = TRUE)

mkdir("foo/bar")
}

}
\seealso{
\code{\link[base:files2]{base::dir.exists()}} and \code{\link[base:files2]{base::dir.create()}} for testing existence and creation of directories.
}
