% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nir.r
\name{nir}
\alias{nir}
\title{Nice rectangulate}
\usage{
nir(x, row = NULL, col = NULL)
}
\arguments{
\item{x}{a either a numeric vector of size one specifying the number of elements
or an object, typically a list, with a length}

\item{row, col}{the number of rows and columns respectively}
}
\value{
a vector of size two specifying the number of rows and column the elements can fix in.
}
\description{
Returns a number of rows and columns that will fit the number of elements in \code{x}.
}
\details{
The main purpose of this function is to find out how many columns or rows are required
to fit a certain number of elements. This is useful when plotting multiple facet
plots with \code{mfrow} and \code{mcol} (see \code{\link[graphics:par]{graphics::par()}}).

If neither \code{row} or \code{col} are specified, squared dimensions are returned.
Otherwise, if \code{row} or \code{col} are specified, rectangular dimensions are returned with
fixed pre-specified dimension.
}
\examples{
nir(11, col = 3) # c(4, 3)
nir(11, row = 3) # c(3, 4)
nir(11) # c(4, 4)

# plot each Iris species as separate facet
# with 2 columns per row
d = split(iris, ~ Species)
par(mfrow = nir(d, col = 2))
Map(
    d, names(d),
    f = \(x, name) plot(Sepal.Width ~ Sepal.Length, data = x, main = name)
    ) |> invisible()

}
