% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize}
\alias{normalize}
\alias{standardize}
\title{Normalize data}
\usage{
normalize(x, select, where, FUN = `-`, STATS = mean, SPLIT = NULL, ...)

standardize(x, select, where, SPLIT = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{select}{a subset of columns that is normalized}

\item{where}{a logical expression indicating rows from which is the statistics calculated}

\item{FUN}{a vectorized function of two arguments representing the application of statistics
(second argument) on the data (first argument), such as \code{-} or \code{/} for centering or scaling.}

\item{STATS}{a function that calculate single statistics
or alternatively a numeric vector with pre-calculated statistics for every selected
column, the vector is recycled as required}

\item{SPLIT}{a factor or a formula passed to \code{\link[base:split]{base::split()}} that further splits \code{x}
into independently normalized subsets, see \code{\link[base:split]{base::split()}}.}

\item{...}{further arguments passed to \code{STATS}}

\item{center}{perform centering}

\item{scale}{perform scaling}
}
\value{
normalized data.frame
}
\description{
Perform data normalization by adjusting selected columns using a calculated statistics.
The statistics can be calculated only on a part of data.
}
\details{
These functions are expanded and user-friendly variant of \code{\link[base:scale]{base::scale()}} for \code{data.frames}.

\code{standardize} is a user-friendly \code{\link[base:scale]{base::scale()}} for \code{data.frames}.
It performs scaling and centering on selected columns against a subset specified by \code{where}.

\code{normalize} is a more general version that allows specifying given statistics
such as \code{\link[base:mean]{base::mean()}} for centering or \code{\link[stats:sd]{stats::sd()}} for scaling,
and operation such as subtraction or division respectively.
Like \code{standardize}, the statistics can be calculated only from specified subset of data
specified by \code{where}.

Data normalization is a general term with a range of meaning. In the most general sense
data normalization is adjusting data, i.e., applying some function using calculated statistic
either provided or more commonly calculated from the data itself.
For instance, \emph{centering} removes mean and \emph{scalig} divides by standard deviation.

A common problem might be to perform normalization against only part of data
such as normalizing your treatments against control and obtaining relative effect sizes.
}
\examples{
normalize(iris, Sepal.Length)
normalize(iris, "Sepal.Length")
normalize(iris, c("Sepal.Length", "Sepal.Width"))

# Normalize all iris columns except species against setosa
normalize(iris, setdiff(names(iris), "Species"), Species == "setosa")

# Standardize all iris column except species against setosa
standardize(iris, setdiff(names(iris), "Species"), Species == "setosa")

# Normalize Sepal.Length by dividing by "setosa" mean
normalize(iris, "Sepal.Length", Species == "setosa", FUN = `/`)

# Normalize each species independently
normalize(iris, setdiff(names(iris), "Species"), SPLIT = ~ Species)

}
\seealso{
See \code{\link[base:scale]{base::scale()}} for standardizing (centering and scaling),
\code{\link[base:sweep]{base::sweep()}} that allows centering and scaling with arbitrary statistics,
\code{\link[base:split]{base::split()}} for dividing data into subset,
and \code{\link[base:subset]{base::subset()}} for selecting subset of data.
}
