% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{operators}
\alias{operators}
\alias{Not}
\alias{\%nin\%}
\title{Operators}
\usage{
Not()

x \%nin\% y
}
\arguments{
\item{x, y}{arguments to operators}
}
\value{
return value depends on the operator, see details
}
\description{
A collection of operators.
}
\section{Details}{
\code{Not()} is a function alias for \code{!} allowing for negation to be used in pipes
and enhance readability in some contexts.

\verb{\%nin\%} is a simple negation of \code{\%in\%}.
}

\examples{
# Operators can't be used in pipes.
# Won't work:
# TRUE |> !
# TRUE |> `!`()

# Works:
TRUE |> Not()

# Following are identical
!("Dinosaur" \%in\% names(iris))
"Dinosaur" \%nin\% names(iris)
}
