% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.r
\name{pad}
\alias{pad}
\title{Pad a vector}
\usage{
pad(x, length, value = NA, side = c("left", "right"), truncate = TRUE)
}
\arguments{
\item{x}{a vector}

\item{length}{a desired length of output vector, if \code{length} is smaller than length of \code{x},
the resulting vector is unchanged.}

\item{value}{a value used for padding the vector}

\item{side}{direction in which the vector will be padded,
either \code{left} where the values are added at the start of the vector,
or \code{right} where the values are added at the end of the vector}

\item{truncate}{if \code{TRUE} (default), truncate the vector to a given length, see details.}
}
\value{
a vector of specified \code{length}
}
\description{
This function will par (extend) a vector with selected values to desired length.
}
\details{
If \code{truncate = TRUE} (default), the vector \code{x} is truncated to guarantee that the returned
vector size is exactly \code{length}. This is done in standard way by taking the first \code{length}
values of the vector \code{x}.
}
\examples{

x = 1:5
pad(x, 10)
pad(x, 10, side = "right")
pad(x, 3) # truncated
pad(x, 3, truncate = TRUE) # unchanged

# type conversion:
pad(x, 10) # integer
pad(x, 10, "foo") # character

# converted even if no padding is done
pad(x, 3, "foo") # character
pad(x, 3, "foo", truncate = FALSE) # character

}
\seealso{
\code{\link[=strpad]{strpad()}} for padding strings
}
