% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace2.r
\name{replace2}
\alias{replace2}
\alias{replace2.default}
\alias{replace2.data.frame}
\title{Replace values}
\usage{
replace2(x, values, replace, ...)

\method{replace2}{default}(x, values, replace, ...)

\method{replace2}{data.frame}(x, values, replace, select = NULL, ...)
}
\arguments{
\item{x}{a subsetable object, such as a vector, matrix, or a data.frame}

\item{values}{a vector of values in 'x' to be replaced by values in \code{replace}}

\item{replace}{replacement values}

\item{...}{further arguments passed to or from other methods}

\item{select}{a numeric or a character vector specifying which columns are selected
for replacement}
}
\value{
an object with replaced values
}
\description{
Replace the values in \code{x}, specified by the vector of values
\code{values} with the values in the vector \code{replace}.
}
\details{
This is a variant of \code{\link[base:replace]{base::replace()}} that might be more useful when replacing
several differet values, e.g., when cleaning and recoding data.
}
\examples{
vec = c("foo", "bar", "baz")
replace2(vec, "bar", "spam")

# No error is reported if values are not in object
replace2(vec, "spam", "eggs")

# Useful when recoding
replace2(
    iris,
    c("setosa", "versicolor", "virginica"),
    c("set", "ver", "vir")
    )


}
\seealso{
\code{\link[base:replace]{base::replace()}} for a different interface for a similar idea,
\link[base:Extract]{base::Extract} for general subsetting and replacement

Consider also \code{\link[base:transform]{base::transform()}}, \code{\link[base:with]{base::with()}}, and \code{\link[base:with]{base::within()}} for
\code{data.frame} modification.
}
