% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rntc.r
\name{rntc}
\alias{rntc}
\title{Row names to column}
\usage{
rntc(x, name = "name", remove = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{name}{a character string specifying a name for the newly created column}

\item{remove}{logical, remove column names}
}
\value{
a data.frame with row names transformed into the first column
}
\description{
Transform row names of a \code{data.frame} to first column.
}
\examples{
rntc(mtcars)

# use "model" as a more appropriate label
rntc(mtcars, "model")

# do not remove old row names
rntc(mtcars, remove = FALSE)

}
\seealso{
\code{\link[base:colnames]{base::rownames()}} for setting or resetting row names,
or see the \code{row.names} argument in \code{\link[base:data.frame]{base::data.frame()}} for the construction
of \code{data.frames} with the rownames attribute.
}
