% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.r
\name{setNames}
\alias{setNames}
\alias{set_names}
\title{Set the names of an object}
\usage{
setNames(object = nm, nm, dim = NULL)

set_names(object = nm, nm, dim = NULL)
}
\arguments{
\item{object}{an object for which generics \code{\link[base:names]{base::names<-()}} or \code{\link[base:dimnames]{base::dimnames<-()}} exist}

\item{nm}{a character vector of names to assign to the object}

\item{dim}{an integer, dimension to be changed}
}
\description{
This is a convenience function that sets the names of an object
and returns the object. It is a direct replacement of \code{\link[stats:setNames]{stats::setNames()}}
with an additional parameter specifying the dimension.
This allows for setting row names for \code{data.frame} as well as modifying
dimnames of matrices and arrays.
}
\details{
By default, the function is a direct replacement of \code{\link[stats:setNames]{stats::setNames()}}
as a functional version of the \verb{names<-()}
However, if the \code{dim} is specified \code{setNames()} will instead replace the \code{dimnames}
of the object specified by \code{dim}. See examples.

The side-effect of \code{\link[base:dimnames]{base::dimnames()}} is a conversion of all dimnames into a character vector.
}
\examples{

setNames(c(1, 2, 3), c("a", "b", "c"))
# Useful with pipes
c(1, 2, 3) |> setNames(c("a", "b", "c"))

# alias
c(1, 2, 3) |> set_names(c("a", "b", "c"))

# Setting row of matrix and data.frame
matrix(1:4, 2, 2) |>
   setNames(c("a", "b"), 1)
data.frame("foo" = 1:2, "bar" = 3:4) |>
   setNames(c("a", "b"), 1)

}
\seealso{
\code{\link[stats:setNames]{stats::setNames()}} for the original function,
\code{\link[base:names]{base::names()}}, \code{\link[base:dimnames]{base::dimnames()}} for modifying names and dimnames
}
