% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.r
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Sort data frame}
\usage{
\method{sort}{data.frame}(x, decreasing = FALSE, by = 1, ...)
}
\arguments{
\item{x}{a data.frame}

\item{decreasing}{logical vector passed to \code{\link[base:order]{base::order()}} indicating increasing or decreasing
order of a particular column is used, recycled as necessary}

\item{by}{column names or indices, by default the first column is used}

\item{...}{arguments passed to \code{\link[base:order]{base::order()}} such as \code{method} or \code{na.last}}
}
\value{
a data.frame sorted according to selected columns
}
\description{
Sort data frame according to selected columns.
}
\details{
This is an S3 method for the \code{\link[base:sort]{base::sort()}} generics and in many way
it is similar to the later \code{\link[base:sort_by]{base::sort_by()}} function with the major difference
is the interface, while \code{\link[base:sort_by]{base::sort_by()}} is using a formula interface and thus
non-standard evaluation, this function is expecting index or column names.

The consequence is that column names for sorting can be provided with a variable,
but we are limited to the parameter order of \code{\link[base:sort]{base::sort()}}, with \code{decreasing} preceding
\code{by}.
}
\examples{
# By default, first column is used
sort(iris)

# Following are identical
sort(iris, by = "Species")
iris |> sort(by = "Species")
iris[order(iris[["Species"]]),]

# Pass columns using a variable
col = "Species"
iris |> sort(by = col)

# With multiple columns, order is important
iris |> sort(by = c("Sepal.Length", "Sepal.Width"))
iris |> sort(by = c("Sepal.Width", "Sepal.Length"))

# Compare with 'sort_by'
iris |> sort_by( ~ Species)

# Won't work:
# iris |> sort_by("Species")
# iris |> sort_by(col)

}
\seealso{
The \code{\link[base:sort]{base::sort()}} is the generic function.
\code{\link[base:order]{base::order()}} is internally called to determines ordering.
\code{\link[base:sort_by]{base::sort_by()}} for an alternative interface with a formula allowing non-standard evaluation
}
