% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.r
\name{strpad}
\alias{strpad}
\title{Pad a character vector}
\usage{
strpad(
  x,
  width = NULL,
  value = " ",
  side = c("left", "right"),
  truncate = TRUE
)
}
\arguments{
\item{x}{a character vector}

\item{width}{a desired width, by default, the width of the largest string is used
(\code{max(nchar(x))})}

\item{value}{a string, typically a single character, longer strings are allowed, but
won't produce desired width.}

\item{side}{direction in which the strings will be padded,
either \code{left} where the padding is added at the start of the string,
or \code{right} where the padding is added at the end of the string}

\item{truncate}{if \code{TRUE} (default), truncate the strings to a given length, see details.}
}
\value{
a character vector of strings with specified \code{width}
}
\description{
Pad a character vector to the same width (number of characters).
}
\details{
If \code{truncate = TRUE} (default), the strings in \code{x} are truncated to the desired width
using \code{substring(x, 1, width)}.
}
\examples{
x = c("a", "aa", "aaa")
strpad(x)
strpad(x, 5)
strpad(x, side = "right")
strpad(x, value = "_")

# truncation
strpad(x, 1)
strpad(x, 1, truncate = FALSE)


}
\seealso{
\code{\link[=pad]{pad()}} for padding vectors
\code{\link[base:sprintf]{base::sprintf()}} for C-style string formatting
\code{\link[base:format]{base::format()}} and \code{\link[base:formatc]{base::formatC()}} for a more general formatting functions
}
