% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varclust.r
\name{varclust}
\alias{varclust}
\alias{varcut}
\alias{plot.varclust}
\title{Variable Clustering}
\usage{
varclust(
  x,
  similarity = c("pearson", "kendall", "spearman"),
  method = "complete",
  use = "na.or.complete"
)

varcut(x, value = 0.7)

\method{plot}{varclust}(
  x,
  ylab = "Similarity",
  main = "",
  at = c("range", "values"),
  digits = 1,
  value = 0.7,
  val.col = "red",
  val.lwd = 3,
  val.lty = 1,
  ...
)
}
\arguments{
\item{x}{for \code{varclust()} a data.frame, categorical variables should be provided as factors
and are dummy-coded, for \code{varcut()} or \code{plot.varclust()} an object of class \code{varclust}}

\item{similarity}{a character string specifiying the chosen method for comparing variables,
currently only correlation methods are supported}

\item{method}{a character string specifying the linking for the hierarchical clustering,
the default "complete" makes the most sense as it creates continuous clusters}

\item{use}{a character string giving a method for computing correlation
in the presence of missing variables, see \code{\link[stats:cor]{stats::cor()}} for details}

\item{value}{a single numeric value giving the similarity threshold at which the tree
from hierarchical clustering is cut to define clusters}

\item{ylab, main, at, digits}{graphical options for plot}

\item{val.col, val.lwd, val.lty}{graphical options for the threshold value line}

\item{...}{other graphical options passed to \code{\link[stats:dendrogram]{stats::plot.dendrogram()}}}
}
\value{
for \code{varclust()} an object of the class \code{varclust} containing the calculated similarity matrix
and hierarchical clustering.
for \code{varcut()} a list with diagnostic information and selected variable names
}
\description{
Perform variable clustering by calculating a similarity measure and then perform
hierarchical clustering to build a similarity tree of variables.
}
\details{
Variable clustering is a simple method for feature selection, or at least dimensional
reduction for highly correlated variable spaces. It is also a highly-effective method
for visualizing correlation matrices with the ability to reduce large matrices to a small
clusters of related variables.

\code{varclust()} performs variable clustering using correlation and a hierarchical clustering.

\code{varcut()} cuts the tree from hierarchical clustering at some similarity threshold,
this defines clusters of highly related variables.
A single variable is selected out of each cluster, peforming a feature selection.

\code{plot.varclust()} is a plotting method
}
\examples{
v = varclust(iris)
varcut(v)$select

}
