% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2col.r
\name{vec2col}
\alias{vec2col}
\title{Convert a vector to colors}
\usage{
vec2col(
  x,
  palette,
  type = c("hcl", "palette", "colorRamp", "colors", "colours"),
  ...
)
}
\arguments{
\item{x}{an input vector with preferably ordinal or categorical values}

\item{palette}{a name of a palette, either a name (for type \code{hcl} or \code{palette}),
or a vector of colours (for type \code{colorRamp}, \code{colors} or \code{colours}).}

\item{type}{type of palette, see details}

\item{...}{other parameters passed to the palette functions}
}
\value{
a named vector of colours of the same length as \code{x}
}
\description{
Generate a vector of colours where each unique element of the input vector will be assigned
a colour value.
}
\details{
If the \code{x} is a factor, then levels are used for categories.
This allows consistent colour generation across subsets.
Otherwise, sorted unique elements are used instead.

The parameters \code{palette} and \code{type} determine which colour generating method is used.
The \code{hcl}, \code{palette}, and \code{colorRamp} call \code{\link[grDevices:palettes]{grDevices::hcl.colors()}},
\code{\link[grDevices:palette]{grDevices::palette.colors()}}, and \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}} respectively with
\code{palette} as its first argument specifying the name of the palette (for \code{hcl} and \code{palette})
or the vector of colours from which palette is generated (for \code{colorRamp}).
If the \code{type} is either \code{color} or \code{colours}, the \code{palette} is expected to be a vector
of colours which is then directly used. The vector is recycled as required.
}
\examples{
# Convenient way to generate colours for dataset
vec2col(iris$Species, "Roma")

# For < 8 categories, `palette` is more divergent
vec2col(iris$Species, "Set 2", "palette")

# When Passing factor, levels are used
mtcars$carb |> unique() # no 5 or 7
carb = factor(mtcars$carb, levels = 1:8)
vec2col(carb, "Set 2", "palette")

# E.g., smooth colour transition despite incomplete representation
col = vec2col(carb, c("blue", "red"), "colorRamp")
plot(mtcars$cyl, mtcars$carb, col = col, pch = 19, cex = 2)

}
\seealso{
\code{\link[grDevices:palettes]{grDevices::hcl.colors()}}, \code{\link[grDevices:palette]{grDevices::palette.colors()}}, \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}
for colour-generating functions used internally.
}
