% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.r
\name{divide}
\alias{divide}
\alias{divide.default}
\alias{divide.data.frame}
\alias{divide.matrix}
\title{Divide an object into equal parts}
\usage{
divide(x, k, ...)

\method{divide}{default}(x, k, ...)

\method{divide}{data.frame}(x, k, dim = 1, ...)

\method{divide}{matrix}(x, k, dim = 1, ...)
}
\arguments{
\item{x}{an object to be split}

\item{k}{number of objects}

\item{...}{arguments to be passed to or from methods}

\item{dim}{a dimension along which the division will be made, e.g.,
for \code{matrix} and \code{data.frame} 1 (default) divides along rows and 2 along columns}
}
\value{
a list of \code{k} objects derived from \code{x}, see details
}
\description{
Divide an object such as vectors, lists, matrices or data.frames into an \code{k} number of parts.
}
\details{
If the object of size \code{n} is divisible by \code{k} such that \code{n \%\% k = 0},
the object is split into \code{k} parts of equal size of size \code{n \%/\% k}.
Otherwise, the first \code{n \%\% k} parts will have size \code{n \%/\% k + 1}.

The object inheriting from \code{matrix} or \code{data.frame} can be divided along their
rows (\code{dim = 1}) or columns (\code{dim = 2}).
Arrays are currently unsupported and treated as vectors.
}
\examples{
l = letters[1:10]

# perfect division
divide(l, 2)
divide(l, 5)

# unequal division
divide(l, 3)
divide(l, 4)

# list are vectors
divide(as.list(l), 3)

# matrices and data.frames
divide(iris, 3)
divide(iris, 3, 2)
divide(as.matrix(iris), 3)

arr = array(letters, c(2,2,2)) # array 2x2x2
identical(
  divide(arr, 3),
  divide(as.vector(arr), 3)
  )

}
\seealso{
\code{\link[base:split]{base::split()}} for spliting an object according to factors
}
