% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.r
\name{select}
\alias{select}
\alias{deselect}
\alias{fill}
\title{Select and deselect elements}
\usage{
select(x, i, drop = TRUE)

deselect(x, i)

fill(x, i, drop = TRUE)
}
\arguments{
\item{x}{a vector, list, data.frame, or other object for which \code{[} and \code{[[} are defined.}

\item{i}{a numeric index or a character vector if \code{x} is named;
an error is reported if the index \code{i} is out of bound or the name is not in the vector}

\item{drop}{drop extra dimensions, relevant only for certain objects and only for \code{select},
e.g., if \code{x} is a data.frame and \code{length(x) == 1}, a vector is returned as if \code{[[} was called.
\code{deselect} should never drop dimensions, if you want to deselect all but one columns,
just select that column.}
}
\value{
an array-like object derived by subsetting \code{x}
}
\description{
Pipe-friendly subsetting functions for vectors, lists, and data.frames.
}
\details{
\code{select()} will subset a \code{vector}, \code{list} or a \code{data.frame} based on character or
numeric indices \code{i}. In many ways it behaves like \code{[}, but it is safer in the sense
that non-existing indices will always throw error (\code{[} returns \code{NA} instead)

\code{deselect()} is the opposite of \code{select()}, it removes elements specified by \code{i}
from the object. Non-existing elements are ignored.

\code{fill()} is a variant of \code{select()} that works with non-existing elements as well.
If \code{i} is in \code{x}, it the output is identical to \code{select()}, otherwise it fills
the missing elements with \code{NA} (for vectors or data.frames) or \code{NULL} (for lists).
}
\examples{
select(iris, "Species")

# Compare:
iris |> getElement("Species")
iris |> (\(x) x[["Species"]])()
iris |> select("Species")

# For multiple elements/columns
iris |> (\(x) x[c("Sepal.Width", "Sepal.Length")])()
iris |> select(c("Sepal.Width", "Sepal.Length"))

# Won't work:
# iris |> (\(x) x[[c("Sepal.Width", "Sepal.Length")]])()
# iris |> getElement(c("Sepal.Width", "Sepal.Length"))

# No non-standard evaluation
try(iris |> select(Species))
columns = c("Sepal.Width", "Sepal.Length")
iris |> select(columns)

# Don't drop columns:
iris |> select("Species", drop = FALSE)

# Remove columns
iris |> deselect("Species")
iris |> deselect(c("Sepal.Width", "Sepal.Length"))

}
\seealso{
\link[base:Extract]{base::Extract} and \code{\link[base:subset]{base::subset()}}
}
